+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.7.0 (Unreleased)
 
-* Added methods `OrgVdcNetwork.Update`, `OrgVdcNetwork.UpdateAsync`, and `OrgVdcNetwork.Rename`
-* Added methods `EdgeGateway.Update` and `EdgeGateway.UpdateAsync`
+* Added methods `OrgVdcNetwork.Update`, `OrgVdcNetwork.UpdateAsync`, and `OrgVdcNetwork.Rename` [#292](https://github.com/vmware/go-vcloud-director/pull/292)
+* Added methods `EdgeGateway.Update` and `EdgeGateway.UpdateAsync` [#292](https://github.com/vmware/go-vcloud-director/pull/292)
 
 ## 2.6.0 (March 13, 2010)